requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-identity
azure-servicebus
pyjwt[crypto]
psycopg2-binary
pydantic
requests
python-dotenv
psycopg[pool]